Move bfcache invalidation logic to separate file
